[API] remove unused use case
